LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-output-element/mutations.window-expected.txt


FAIL Descendant mutations and output.value and .defaultValue assert_equals: .defaultValue after setting textContent expected "x" but got ""
PASS output and output.form.reset()
